- Summary: fix bug137880:Import attached design file, preview results different in html and PDF. - Bugzilla Bug (s) Resolved: 137880 - Description: fixed by consider device scale during 3D engine initialization. - Tests Description  : Manual test - Files Edited: @changelist - Files Added: @addlist - Files Deleted: @deletelist - Notes to Build Team:
- Notes to Developers:

- Notes to QA:

- Notes to Documentation:

+++ chart/org.eclipse.birt.chart.engine/src/org/eclipse/birt/chart/computation/withaxes/PlotWith3DAxes.java
@@ -2120,7 +2120,7 @@ public class PlotWith3DAxes extends PlotWithAxes
 				}
 				else if ( isZeroValue == true )
 				{
-					percentileValue = new Double ( (double)1.0 / iOrthogonalCount );
+					percentileValue = new Double ( 1d / iOrthogonalCount );
 				}
 
 				dpa[i] = new DataPointHints( oDataBase,
